fix: address code review issues in icon extraction PR
- Rename Icons.tsx to icons.tsx (kebab-case file naming convention)
- Merge duplicate StrokeIconProps into single IconProps interface
- Add className prop to IconProps for styling flexibility
- Add optional aria-label prop with conditional role="img"/aria-hidden
- Add missing aria-label to Cards/Explorer layout toggle buttons
